Refactor weather & implement in CLI
The weather code was created by user Elefantex, this commit moves and cleans up the functions so they're D.R.Y. and can be used elsewhere, and removes the need for a hardcoded API key.

This commit uses these refactored functions to make them callable from the CLI, and makes the GUI use the CLI interface for weather, in line with the rest of GUI.
Keeping this all in one commit partially to not obscure Elefantex's original authorship.