Create complete GUI installer for Windows
The build script automatically sets up the venv and required GUI packages, and creates shortcuts to the users desktop and start menu applications folde, making it searchable.

It doesn't use PyInstaller, instead it creates shortcuts that open up a new (optionally hidden) PowerShell window, which enables the venv and starts the gui normally.